feat: search for queries and display results
Add react components for the following:
* Search Page - top level search page component
* Search Bar - for entering queries
* Search Results - list of books returned from query

In the event of an invalid query, the search results are cleared.

Navigation is handled by React Router, so back and forward work with the
search page.

This completes rubric item "Does the search page have a search input
that lets users search for books?"