fix: remove duplicates when changing from search
If a book is already on a user's shelf, and the user searches for the
book and changes the shelf from within the search page, the main page
would show the book on multiple shelves. This happens because the search
result doesn't always have the "shelf" property set. My solution was to
search the users shelves for books with the same title and changes the
shelf manually in this case.